- change 'refreshfreq' from int to double.
A double is already used in CalcSmoothRatio. Further to this, displays that use TV standard frequencies do not use perfect 60Hz frequencies, but frequencies such as 59.94Hz, 23.976Hz (precisely, 24*(1000/1001) = 23.9760239760239760...) etc.

Reference: 'https://www.ghacks.net/2010/04/28/59-hertz-refresh-rate/', 'http://www.paradiso-design.net/videostandards_en.html'.

+++ source/platform/posix/cocoa/i_main.mm
@@ -139,7 +139,7 @@ void I_DetectOS()
 
 FArgs* Args; // command line arguments
 
-int refreshfreq;
+double refreshfreq;
 bool batchrun;
